Final improvements to styling before release announcement
I changed the chart type selector to be radio buttons (Bar, Line, Scatter)
which breaks up the UI. I also added a "Charting options" <h3> title and
tweaked the display of the "Swap X and Y" button.

The design is now good enough that I'm happy to announce it. Closes #3